Add override mechanism.  Simply call TWN "Taiwan".
Background:

Using this Gem on our e-commerce site, we've received complaints from
users in Taiwan that calling it "Taiwan, Province of China" is
politically insensitive / inflammatory.  In our codebase we're
special-casing it right now, but perhaps other users of this Gem might
not be aware of this issue either.  Therefore see attached a patch to
introduce an override mechanism, which allows the country list to
remain auto-generated, but provides a simple override system.  This
might also be useful for those who want to refer to USA as simply
"United States", for example.

Also, add a patch to simply refer to TWN as "Taiwan".

* lib/iso_country_codes/iso_3166_1.rb: Change TWN name to "Taiwan"
  only.
* README.rdoc: Reflect new override feature in README.
* overrides.yml: Add file.  Introduce "Taiwan" exception.
* rakelib/iso_3166_1.rb: Apply the overrides after scraping Wikipedia.